Move dummy_attention_generator to its own file for deduplication
