+++ prompts/generator_builder.py
@@ -14,22 +14,14 @@ from dynamicprompts.generators import (
 
 logger = logging.getLogger(__name__)
 
-class DummyAttentionGenerator:
-    def __init__(self, generator: PromptGenerator, *args, **kwargs):
-        self._generator = generator
-        logger.warning("Using the DummyAttentionGenerator, this is possibly because you don't have spacy installed. Install it with `python -m pip install dynamicprompts[attentiongrabber]`")
-
-    def generate(self, *args, **kwargs) -> List[str]:
-        return self._generator.generate(*args, **kwargs)
-
 from dynamicprompts.generators.magicprompt import MagicPromptGenerator
 
 try:
     from dynamicprompts.generators.attentiongenerator import AttentionGenerator
 except NameError:
-    AttentionGenerator = DummyAttentionGenerator
-
-logger = logging.getLogger(__name__)
+    from prompts.dummy_attention_generator import (
+        DummyAttentionGenerator as AttentionGenerator,
+    )
 
 DEFAULT_MAGIC_MODEL = "Gustavosta/MagicPrompt-Stable-Diffusion"
